Global error handling

+++ src/controller/product.controller.js
@@ -1,48 +1,37 @@
 const Products = require("../../model/Products");
 const Carts = require("../../model/Cart");
 const fs = require("fs");
+const {verifyRequired, isExistingUser, fetchUser } = require("../helper/validation")
+const {
+  handleSuccess,
+  handleForbidden,
+  handleNotFound,
+  handleBadRequest,
+  handleEmptyBody
+} = require("../../middleware/errorHandling");
 
 const addproduct = async (req, res) => {
   if (Object.keys(req.body).length == 0) {
-    return res.status(500).send({
-      status: 500,
-      message: "Empty body",
-    });
+    return handleEmptyBody(res);
   } else {
     let product = req.body;
     if (!req.file?.filename) {
-      return res.status(400).send({
-        status: 400,
-        message: "Product Image is required",
-      });
+     return handleNotFound(res, "Product Image Not found");
     }
-    if (
-      !req.body.price ||
-      !req.body.name ||
-      !req.body.category ||
-      !req.body.stock
-    ) {
-      return res.status(400).send({
-        status: 400,
-        message: "Product Name, Price,stock,category are required",
-      });
+    let errors = verifyRequired(req, ["name", "price","stock","category"])
+    if (errors.length != 0) {
+      return handleNotFound(res, errors);
     }
     try {
       product.image = req.file.filename;
       const newProduct = new Products(product);
       await newProduct.save();
       //Product.create(product)
-      return res.status(201).send({
-        status: 201,
-        message: "Product Added",
-      });
+      return handleSuccess(res, "Product Added");
     } catch (error) {
       //Removeing file if product is not inserted
       fs.unlinkSync(`public/assets/images/${req.file.filename}`);
-      return res.status(400).send({
-        status: 400,
-        message: "Something went wrong",
-      });
+      return handleBadRequest(res, "Something went wrong");
     }
   }
 };
@@ -53,19 +42,12 @@ const getbycategory = async (req, res) => {
   try {
     const products = await Products.find({ category: regex });
     if (products.length === 0) {
-      return res
-        .status(404)
-        .send({ status: 404, message: "Category Not Found" });
+      return handleNotFound(res, "Category Not Found");
     } else {
-      return res
-        .status(200)
-        .send({ status: 200, message: "success", data: products });
+      return handleSuccess(res,products);
     }
   } catch (error) {
-    return res.status(400).send({
-      status: 400,
-      message: "Something went wrong",
-    });
+    return handleBadRequest(res);
   }
 };
 
@@ -73,165 +55,106 @@ const getbestseller = async (req, res) => {
   try {
     const products = await Products.find({ bestSeller: true });
     if (products.length === 0) {
-      return res
-        .status(404)
-        .send({ status: 404, message: "Category Not Found" });
+      return handleNotFound(res, "Category Not Found");
     } else {
-      return res
-        .status(200)
-        .send({ status: 200, message: "success", data: products });
+      return handleSuccess(res,products);
     }
   } catch (error) {
-    return res.status(400).send({
-      status: 400,
-      message: "Something went wrong",
-    });
+    return handleBadRequest(res);
   }
 };
 
 const deleteProduct = async (req, res) => {
   try {
-    let isdelete = await Products.findByIdAndDelete(req.params.id)
+    let isdelete = await Products.findByIdAndDelete(req.params.id);
     if (isdelete) {
       //removing image
-      fs.unlinkSync(`public/assets/images/${isdelete.image}`)
-      return res.status(200).send({
-        status: 200,
-        message: "Product deleted"
-      })
+      fs.unlinkSync(`public/assets/images/${isdelete.image}`);
+      return handleSuccess(res, "Product Deleted");
     } else {
-      return res.status(404).send({
-        status: 404,
-        message: "Record not found"
-      })
+      return handleNotFound(res, "Product Not Found");
     }
+  } catch (error) {
+    return handleBadRequest(res, "Something went wrong");
   }
-  catch (e) {
-    return res.status(400).send({
-      status: 400,
-      message: "Something went wrong"
-    })
-  }
-}
+};
 
 const addtocart = async (req, res) => {
-  user_id = req.user.id
-  const { product_id } = req.body
+  user_id = req.user.id;
+  const { product_id } = req.body;
   // return res.send({user_id,product_id})
   if (!product_id) {
-    return res.status(404).send({
-      status: 404,
-      message: 'Product Id is required'
-    })
+    return handleNotFound(res, "Product id is Required");
   }
-  alreadyInserted = await Carts.count({ user_id, product_id })
+  alreadyInserted = await Carts.count({ user_id, product_id });
   if (alreadyInserted != 0) {
-    return res.status(403).send({
-      status: 403,
-      message: "Already inserted"
-    })
+    return handleForbidden(res, "Already inserted");
   }
   try {
     let newCart = new Carts({
       user_id,
       product_id,
-      qty: 1
-    })
-    await newCart.save()
-    return res.status(201).send({
-      status: 201,
-      message: 'Product added'
-    })
+      qty: 1,
+    });
+    await newCart.save();
+    return handleSuccess(res, "Product inserted");
+  } catch (error) {
+    return handleBadRequest(res);
   }
-  catch (error) {
-    return res.status(400).send({
-      status: 400,
-      message: 'Unable to add product to cart',
-    })
-  }
-}
+};
 
 const getUserCart = async (req, res) => {
   try {
-    user_id = req.user.id
-    let UserProduct = await Carts.find({ user_id }).populate({ path: 'product_id', options: { strictPopulate: false } })
-    return res.send({
-      status: 200,
-      message: "Success",
-      data: UserProduct
+    user_id = req.user.id;
+    let UserProduct = await Carts.find({ user_id }).populate({
+      path: "product_id",
+      options: { strictPopulate: false },
     });
+    return handleSuccess(res,UserProduct);
   } catch (error) {
-    return res.send({
-      status: 400,
-      message: "Something went wrong",
-    });
+    return handleBadRequest(res);
   }
-
-}
+};
 
 const removeFromCart = async (req, res) => {
   // const { user_id, product_id } = req.body
-  let user_id = req.user.id
+  let user_id = req.user.id;
   let product_id = req.params.productid;
   if (!product_id) {
-    return res.status(400).send({
-      status:400,
-      message: 'Product Id is required'
-    })
+    return handleNotFound(res, "Product id is required");
   }
   try {
     const result = await Carts.deleteOne({
       user_id: user_id,
-      product_id: product_id
+      product_id: product_id,
     });
     if (result.deletedCount === 0) {
-      return res.status(400).send({
-        status:400,
-        message: "Invalid product_id"
-      })
+      return handleBadRequest(res, "Invalid product id");
     }
-    return res.status(200).send({
-      status:200,
-      message: "Product removed"
-    })
+    return handleSuccess(res, "Product removed");
+  } catch (error) {
+    return handleBadRequest(res);
   }
-  catch (error) {
-    return res.status(400).send({
-      status:400,
-      message: "Something went wrong"
-    })
-  }
-}
+};
 
 const updatecart = (req, res) => {
-  let cart_id = req.params.cartid
+  let cart_id = req.params.cartid;
   if (Object.keys(req.body).length == 0) {
-      return res.status(500).send({ status:500,message: "Empty body" })
+    return handleEmptyBody(res, "Empty Body");
   }
-  Carts.updateOne(
-    { _id: cart_id },
-    { qty: req.body.qty }
-  ).then(result => {
-    console.log(result);
+  Carts.updateOne({ _id: cart_id }, { qty: req.body.qty })
+    .then((result) => {
+      console.log(result);
       if (result.matchedCount > 0) {
-          return res.status(201).send({
-              status: 201,
-              message: "updated"
-          })
+        return handleSuccess(res, "Update");
+      } else {
+        return handleBadRequest(res, "Invalid Product Id");
       }
-      else {
-          return res.status(400).send({
-              status: 400,
-              message: "Invalid cart id"
-          })
-      }
-  }).catch(error => {
-      return res.status(400).send({
-          status: 400,
-          message: "Something went wrong"
-      })
-  })
-}
+    })
+    .catch((error) => {
+      return handleBadRequest(res, "Something went wrong");
+    });
+};
 
 const searchbyname = async (req, res) => {
   const name = req.params.name;
@@ -239,19 +162,12 @@ const searchbyname = async (req, res) => {
   try {
     const results = await Products.find({ name: regex });
     if (results.length === 0) {
-      return res
-        .status(404)
-        .send({ status: 404, message: "Product Not Found" });
+      return handleNotFound(res, "product not found");
     } else {
-      return res
-        .status(200)
-        .send({ status: 200, message: "success", data: results });
+      return handleSuccess(res,results);
     }
   } catch (error) {
-    return res.status(400).send({
-      status: 400,
-      message: "Something went wrong",
-    });
+    return handleBadRequest(res);
   }
 };
 
@@ -264,5 +180,5 @@ module.exports = {
   getUserCart,
   removeFromCart,
   updatecart,
-  searchbyname
+  searchbyname,
 };
